+++ Changelog.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file
         | SymbolHiddenF 	| `◻` `◼`    	|  Hidden files `Shown`/`Hidden` symbol                           	|
         | HideDotfile   	| `False`    	|  Treat .dotfiles as hidden files on Windows                     	|
 
+    + :sparkles: A command to view current configuration (`view_configuration_status_bar_extended` aliased as `StatusBarExtended: view current configuration settings` in Command Palette)
+
   - __Fixed__
     + :beetle: selection updating on each cursor move even when without a `toggle_selection` argument
     + :beetle: update function called 3 times per path change (one extra for each `from statusbarextended import StatusBarExtended`)